convert RC_SPELL to at_observer during readgame.
