+++ src/attributes/attributes.h
@@ -23,7 +23,13 @@ extern "C" {
 #endif
 
     struct attrib_type;
+    struct region;
+    struct faction;
+
     extern void register_attributes(void);
+
+    void set_observer(struct region *r, struct faction *f, int perception, int turns);
+    int get_observer(struct region *r, struct faction *f);
 
 #ifdef __cplusplus
 }
