+++ src/spells.h
@@ -32,9 +32,6 @@ extern "C" {
     void register_magicresistance(void);
     void register_spells(void);
 
-    void set_observer(struct region *r, struct faction *f, int perception);
-    int get_observer(struct region *r, struct faction *f);
-
     int sp_baddreams(castorder * co);
     int sp_gooddreams(castorder * co);
